pyproject.toml
[project]
name = "bingeable"
version = ""
description = ""
authors = [
    {name = "Mikey Clarke", email = "mikey.clarke@me.com"},
]
license = {text = "All rights reserved."}
requires-python = "^3.10"

[tool.mypy]
mypy_path = 'stubs'

[tool.poetry]
name = "bingeable"
version = "0.0.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
masonite = "^4.15"
emoji-country-flag = "^1.3"
redis = "^4.3.4"
python-slugify = "^6.1"
Babel = "^2.10"

[tool.poetry.dev-dependencies]
bandit = "^1.7"
flake8 = "^4.0"
dlint = "^0.12"
flake8-bugbear = "^22.6"
flake8-simplify = "^0.19"
pep8-naming = "^0.13"
flake8-eradicate = "^1.2"
mypy = "^0.971"
types-requests = "^2.28"
types-python-slugify = "^6.1.0"
types-babel = "^2.10"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
